implement model initializers and custom property accessors

+++ ios-documents-objc/ios-documents-objc/Model/NSString+WordCount.m
@@ -10,9 +10,9 @@
 
 @implementation NSString (WordCount)
 
-- (NSUInteger)wordCount {
+- (int)wordCount {
     
-    NSUInteger words = 0;
+    int words = 0;
     
     NSScanner *scanner = [NSScanner scannerWithString: self];
